[#20] Refactor: extension 분리
